webapp/requirements.txt
Flask
requests
python-dotenv
azure-storage-blob
